+++ components/bot_detector/database/api_public/__init__.py
@@ -1,0 +1,14 @@
+"""
+API Public specific database models and helpers.
+"""
+
+from .models import Label, PredictionFeedback, Prediction_v1, Prediction_v2, Player, Report
+
+__all__ = [
+    "Player",
+    "Prediction_v1",
+    "Prediction_v2",
+    "PredictionFeedback",
+    "Report",
+    "Label",
+]
