+++ components/bot_detector/database/api_public/models/__init__.py
@@ -1,0 +1,14 @@
+from .feedback import PredictionFeedback
+from .label import Label
+from .player import Player
+from .prediction import Prediction_v1, Prediction_v2
+from .report import Report
+
+__all__ = [
+    "Player",
+    "Prediction_v1",
+    "Prediction_v2",
+    "PredictionFeedback",
+    "Report",
+    "Label",
+]
