+++ projects/api_public/pyproject.toml
@@ -27,6 +27,7 @@ packages = ["bot_detector"]
 
 [tool.polylith.bricks]
 "../../bases/bot_detector/api_public" = "bot_detector/api_public"
+"../../components/bot_detector/database" = "bot_detector/database"
 "../../components/bot_detector/kafka" = "bot_detector/kafka"
 "../../components/bot_detector/structs" = "bot_detector/structs"
 "../../components/bot_detector/logfmt" = "bot_detector/logfmt"
